Move relations factories to their own module
