+++ network-api/networkapi/wagtailpages/tests/research_hub/test_author_index.py
@@ -5,8 +5,8 @@ from django.utils import translation
 from wagtail_localize import synctree
 
 from networkapi.wagtailpages.factory import profiles as profiles_factory
-from networkapi.wagtailpages.factory import research_hub as research_factory
 from networkapi.wagtailpages.factory.research_hub import detail_page as detail_page_factory
+from networkapi.wagtailpages.factory.research_hub import relations as relations_factory
 from networkapi.wagtailpages.tests.research_hub import base as research_test_base
 from networkapi.wagtailpages.tests.research_hub import utils as research_test_utils
 
@@ -20,7 +20,7 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
             original_publication_date=(research_test_utils.days_ago(n=14)),
         )
         cls.research_profile = profiles_factory.ProfileFactory()
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=cls.detail_page,
             author_profile=cls.research_profile,
         )
@@ -138,7 +138,7 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
         extra_detail_page = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
         )
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=extra_detail_page,
             author_profile=self.research_profile,
         )
@@ -171,7 +171,7 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
         extra_detail_page = detail_page_factory.ResearchDetailPageFactory(
             parent=self.library_page,
         )
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=extra_detail_page,
             author_profile=self.research_profile,
         )
@@ -208,15 +208,15 @@ class TestResearchAuthorIndexPage(research_test_base.ResearchHubTestCase):
             parent=self.library_page,
             original_publication_date=(research_test_utils.days_ago(n=1)),
         )
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=detail_page_1,
             author_profile=self.research_profile,
         )
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=detail_page_2,
             author_profile=self.research_profile,
         )
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=detail_page_3,
             author_profile=self.research_profile,
         )
