+++ network-api/networkapi/wagtailpages/tests/test_profiles.py
@@ -2,8 +2,8 @@ from django import test
 
 from networkapi.wagtailpages.factory import blog as blog_factories
 from networkapi.wagtailpages.factory import profiles as profile_factories
-from networkapi.wagtailpages.factory import research_hub as research_factory
 from networkapi.wagtailpages.factory.research_hub import detail_page as detail_page_factory
+from networkapi.wagtailpages.factory.research_hub import relations as relations_factory
 from networkapi.wagtailpages.pagemodels import profiles as profile_models
 from networkapi.wagtailpages.pagemodels.blog.blog import BlogAuthors
 
@@ -15,7 +15,7 @@ class ProfileTest(test.TestCase):
 
     def test_filter_research_authors(self):
         research_author_profile = profile_factories.ProfileFactory()
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=detail_page_factory.ResearchDetailPageFactory(),
             author_profile=research_author_profile,
         )
@@ -41,11 +41,11 @@ class ProfileTest(test.TestCase):
         """Return research author profile only once"""
 
         research_author_profile = profile_factories.ProfileFactory()
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=detail_page_factory.ResearchDetailPageFactory(),
             author_profile=research_author_profile,
         )
-        research_factory.ResearchAuthorRelationFactory(
+        relations_factory.ResearchAuthorRelationFactory(
             research_detail_page=detail_page_factory.ResearchDetailPageFactory(),
             author_profile=research_author_profile,
         )
